Add files via upload
My first big descriptive analytics project focused on creating a dashboard representing the historic performance of the European Medicines Agency in approving medicines.  The dataset used for this dashboard was downloaded on June 12, 2022 from the European public assessment reports (EPARs) found at www.ema.europa.eu.  The one constraint for this project was to leverage only Excel for the creation of the dashboard.