Add files via upload
I submitted the final project for UNCW's Data Visualization course.  The goal of this project was to find an interesting and publicly available data source.  Analyze the information that is available and tell a story about the data leveraging dashboards and charts from Tableau.  The word document also provides a description of some of the data challenges that I needed to overcome with the dataset.